added main splash screen

+++ src/test/java/TestSuite_Adventure.java
@@ -22,7 +22,7 @@ public class TestSuite_Adventure {
         System.out.println("Tests run: " + testsrun);
         System.out.println("Dungeons Generated: " + DungeonGenerator.runs + " Fails: " + DungeonGenerator.fails);
         System.out.println(((float) DungeonGenerator.fails / DungeonGenerator.runs) * 100 + "% failure rate");
-        System.out.println("For every test run, " + DungeonGenerator.runs / testsrun + " dungeons were generated");
+        System.out.println("For every test run, " + (float) DungeonGenerator.runs / testsrun + " dungeons were generated");
         System.out.println("-----------------------------");
     }
